requirements.txt
python-dotenv
requests
termcolor
streamlit
Pillow